Moved duplicated code to a trait.
